add edit.hlp generation, huffc and fixhelp build with clang (LE)
The huffc and fixhelp commands previously were dependent on int
being 16-bit. So these commands failed when for example the ran under
MacOS.

These tools now compile with clang, but require little endian.

+++ source/helpbox.h
@@ -3,16 +3,17 @@
 #ifndef HELPBOX_H
 #define HELPBOX_H
 
+#include <stdint.h>
 /* --------- linked list of help text collections -------- */
 struct helps {
     char *hname;
 	char *comment;
-    long hptr;
-    int bit;
-    int hheight;
-    int hwidth;
-	int nexthlp;
-	int prevhlp;
+    uint32_t hptr;
+    uint16_t bit;
+    uint16_t hheight;
+    uint16_t hwidth;
+	uint16_t nexthlp;
+	uint16_t prevhlp;
     void *hwnd;
 	char *PrevName;
 	char *NextName;
